encode functions reads **kwargs for more_data flag
TODO: decode(): if more_data is set in headers of received
packet, then send a pull request to the server to send the
rest of the buffer.

TODO: in buffers, only set more_data if data > BUFFER_SIZE.
if (data == BUFFER_SIZE) or (pull request and data > 0), then
set more_data = 0 and flushBuffer().